+++ sfworks/grove/impl/EntityDecl.cxx
@@ -407,7 +407,7 @@ XincludeDecl::XincludeDecl()
 {
     declType_ = xinclude;
     dataType_ = sgml;
-    setName("XI:" + String::number((long int)this));
+    setName("XI:" + String::number((intptr_t)this));
 }
 
 String XincludeDecl::asAnnotationString() const
